+++ himalaya/__init__.py
@@ -7,7 +7,7 @@ from . import scoring
 from . import utils
 from . import viz
 
-__version__ = '0.4.1'
+__version__ = '0.4.2'
 
 __all__ = [
     backend,